Add utils file for generic functions and classes and modifies ClientViewSet to inherit from utils.GenericViewSet
